dali-api/pyproject.toml
[project]
name = "dali-api"
version = "0.1.0"
description = "搭理app Backend API - AI-powered fashion styling assistant"
authors = [
    {name = "Xiaoshaoqian"}
]
readme = "README.md"
requires-python = ">=3.11"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.128.0"}
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = {version = ">=2.0", extras = ["asyncio"]}
asyncpg = "^0.31.0"
alembic = "^1.17.2"
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alibabacloud-dysmsapi20170525 = "^4.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
ruff = "^0.14.10"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]  # B008: Depends() in function defaults is FastAPI pattern

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
